+++ packages/manager/apps/dedicated/client/app/billing/main/billing-main.html
@@ -11,6 +11,9 @@
         <oui-header-tabs-item state="app.account.billing.main.payments"
             ><span data-translate="billing_main_payments_tab"></span
         ></oui-header-tabs-item>
+        <oui-header-tabs-item state="app.account.billing.main.refunds"
+            ><span data-translate="billing_payment_refunds_title"></span
+        ></oui-header-tabs-item>
     </oui-header-tabs>
 </oui-header>
