+++ packages/manager/apps/dedicated/client/app/billing/main/refunds/billing-payment-refunds.routes.js
@@ -1,0 +1,15 @@
+import template from './billing-refunds.html';
+
+export default /* @ngInject */ ($stateProvider) => {
+  const name = 'app.account.billing.main.refunds';
+
+  $stateProvider.state(name, {
+    url: '/refunds',
+    template,
+    controller: 'BillingRefundsController',
+    resolve: {
+      breadcrumb: /* @ngInject */ ($translate) =>
+        $translate.instant('billing_payment_refunds'),
+    },
+  });
+};
