+++ packages/manager/apps/dedicated/client/app/billing/main/translations/Messages_de_DE.json
@@ -2,5 +2,6 @@
   "billing_main_title": "Meine Rechnungen",
   "billing_main_pay_as_you_go_tab": "Pay As You Go",
   "billing_main_bills_tab": "Rechnungen",
-  "billing_main_payments_tab": "Zahlungsverfolgung"
+  "billing_main_payments_tab": "Zahlungsverfolgung",
+  "billing_payment_refunds_title": "Meine Guthaben"
 }
