+++ packages/manager/apps/dedicated/client/app/billing/main/translations/Messages_en_GB.json
@@ -2,5 +2,6 @@
   "billing_main_title": "My bills",
   "billing_main_pay_as_you_go_tab": "Pay As You Go",
   "billing_main_bills_tab": "Invoices",
-  "billing_main_payments_tab": "Payment tracking"
+  "billing_main_payments_tab": "Payment tracking",
+  "billing_payment_refunds_title": "My refunds"
 }
