+++ packages/manager/apps/dedicated/client/app/billing/main/translations/Messages_es_ES.json
@@ -2,5 +2,6 @@
   "billing_main_title": "Mis facturas",
   "billing_main_pay_as_you_go_tab": "Pay-as-you-go",
   "billing_main_bills_tab": "Facturas",
-  "billing_main_payments_tab": "Seguimiento de los pagos"
+  "billing_main_payments_tab": "Seguimiento de los pagos",
+  "billing_payment_refunds_title": "Mis devoluciones"
 }
