+++ packages/manager/apps/dedicated/client/app/billing/main/translations/Messages_es_US.json
@@ -2,5 +2,6 @@
   "billing_main_title": "Facturación",
   "billing_main_pay_as_you_go_tab": "Pay As You Go",
   "billing_main_bills_tab": "Facturas",
-  "billing_main_payments_tab": "Monitoreo de los pagos"
+  "billing_main_payments_tab": "Monitoreo de los pagos",
+  "billing_payment_refunds_title": "Mis devoluciones"
 }
