+++ packages/manager/apps/dedicated/client/app/billing/main/translations/Messages_it_IT.json
@@ -2,5 +2,6 @@
   "billing_main_title": "Le tue fatture",
   "billing_main_pay_as_you_go_tab": "Pay As You Go",
   "billing_main_bills_tab": "Fatture",
-  "billing_main_payments_tab": "Controllo dei pagamenti"
+  "billing_main_payments_tab": "Controllo dei pagamenti",
+  "billing_payment_refunds_title": "Accrediti"
 }
