+++ packages/manager/apps/dedicated/client/app/billing/main/translations/Messages_pl_PL.json
@@ -2,5 +2,6 @@
   "billing_main_title": "Faktury",
   "billing_main_pay_as_you_go_tab": "Pay As You Go",
   "billing_main_bills_tab": "Faktury",
-  "billing_main_payments_tab": "Informacje o płatnościach"
+  "billing_main_payments_tab": "Informacje o płatnościach",
+  "billing_payment_refunds_title": "Moje zwroty"
 }
