+++ packages/manager/apps/dedicated/client/app/billing/main/translations/Messages_pt_PT.json
@@ -2,5 +2,6 @@
   "billing_main_title": "As minhas faturas",
   "billing_main_pay_as_you_go_tab": "Pay As You Go",
   "billing_main_bills_tab": "Facturas",
-  "billing_main_payments_tab": "Pagamentos"
+  "billing_main_payments_tab": "Pagamentos",
+  "billing_payment_refunds_title": "As minhas notas de crédito"
 }
