+++ packages/manager/apps/dedicated/client/app/billing/payment/translations/Messages_de_DE.json
@@ -3,7 +3,6 @@
   "billing_payment_method_title": "Meine Zahlungsarten",
   "billing_payment_ovhaccount_title": "Mein Prepaid-Account",
   "billing_payment_vouchers_title": "Meine Gutscheine",
-  "billing_payment_refunds_title": "Meine Guthaben",
   "billing_payment_fidelity_title": "Mein Treuebereich",
   "billing_payment_credits_title": "Meine Guthaben",
   "billing_payment_transations_title": "Transaktionen"
