+++ packages/manager/apps/dedicated/client/app/billing/payment/translations/Messages_en_GB.json
@@ -3,7 +3,6 @@
   "billing_payment_method_title": "My payment methods",
   "billing_payment_ovhaccount_title": "My prepaid account ",
   "billing_payment_vouchers_title": "My vouchers",
-  "billing_payment_refunds_title": "My refunds",
   "billing_payment_fidelity_title": "My Loyalty control panel",
   "billing_payment_credits_title": "My credits",
   "billing_payment_transations_title": "Transactions"
