+++ packages/manager/apps/dedicated/client/app/billing/payment/translations/Messages_es_ES.json
@@ -3,7 +3,6 @@
   "billing_payment_method_title": "Mis formas de pago",
   "billing_payment_ovhaccount_title": "Mi cuenta de prepago",
   "billing_payment_vouchers_title": "Códigos promocionales",
-  "billing_payment_refunds_title": "Mis devoluciones",
   "billing_payment_fidelity_title": "Área de fidelidad",
   "billing_payment_credits_title": "Mi crédito",
   "billing_payment_transations_title": "Transacciones"
