+++ packages/manager/apps/dedicated/client/app/billing/payment/translations/Messages_fr_FR.json
@@ -3,7 +3,6 @@
   "billing_payment_method_title": "Mes moyens de paiement",
   "billing_payment_ovhaccount_title": "Mon compte prépayé",
   "billing_payment_vouchers_title": "Mes vouchers",
-  "billing_payment_refunds_title": "Mes avoirs",
   "billing_payment_fidelity_title": "Mon espace fidélité",
   "billing_payment_credits_title": "Mes crédits",
   "billing_payment_transations_title": "Transactions"
