+++ packages/manager/apps/dedicated/client/app/billing/payment/translations/Messages_it_IT.json
@@ -3,7 +3,6 @@
   "billing_payment_method_title": "I tuoi metodi di pagamento",
   "billing_payment_ovhaccount_title": "Il tuo conto prepagato",
   "billing_payment_vouchers_title": "I tuoi voucher",
-  "billing_payment_refunds_title": "Accrediti",
   "billing_payment_fidelity_title": "Il tuo spazio fedeltà",
   "billing_payment_credits_title": "I tuoi crediti",
   "billing_payment_transations_title": "Transazioni"
