+++ packages/manager/apps/dedicated/client/app/billing/payment/translations/Messages_pl_PL.json
@@ -3,7 +3,6 @@
   "billing_payment_method_title": "Moje sposoby płatności",
   "billing_payment_ovhaccount_title": "Moje konto przedpłacone",
   "billing_payment_vouchers_title": "Moje kody",
-  "billing_payment_refunds_title": "Moje zwroty",
   "billing_payment_fidelity_title": "Konto prepaid",
   "billing_payment_credits_title": "Moje zasilenia",
   "billing_payment_transations_title": "Transakcje"
