+++ packages/manager/apps/dedicated/client/app/billing/payment/translations/Messages_pt_PT.json
@@ -3,7 +3,6 @@
   "billing_payment_method_title": "Métodos de pagamento",
   "billing_payment_ovhaccount_title": "Conta pré-paga",
   "billing_payment_vouchers_title": "Vouchers",
-  "billing_payment_refunds_title": "As minhas notas de crédito",
   "billing_payment_fidelity_title": "Espaço Fidelização (pontos)",
   "billing_payment_credits_title": "Créditos",
   "billing_payment_transations_title": "Transações"
